Fix error handling in jira_backlog_issue and jira_sprint tables and cleanup general error handling in all tables fixes #61
